Split js into modules
