+++ index.html
@@ -16,7 +16,7 @@
       rel="stylesheet"
     />
     <link rel="stylesheet" href="style.css" />
-    <script defer src="main.js"></script>
+    <script defer type="module" src="main.js"></script>
   </head>
   <body>
     <h1>Mauricio Robayo</h1>
